fix: resolve low-severity code review findings
- L1 (handlers.rs): or_default() instead of or_insert_with(Vec::new);
  notification timestamp now includes seconds (%H:%M:%S)
- L2 (systems.rs): merge duplicate use crate::handlers:: imports into one;
  last_seen/created_date fallback changed from Utc::now() to None so
  never-seen systems no longer appear as "just now"
- L3 (users.rs): field validation moved before transaction open so no
  idle transactions on bad input; role now validated against known values
  (admin/editor/runner/viewer) — arbitrary strings can no longer be stored
- L4 (client.rs): result timestamp uses format("%Y-%m-%dT%H:%M:%S")
  instead of .to_string() which produced a non-standard chrono Debug string
- L5 (scheduler.rs): COUNT(*) columns fetched as i64 (not i32) matching
  SQLite's actual return type — avoids silent overflow on large deployments
- L6 (config.rs): removed duplicate "Config Implementation" comment block
- L7 (tests.rs): merge duplicate use crate::handlers:: imports; add
  HashMap to top-level imports and remove std::collections:: inline references

+++ scmserver/src/scheduler.rs
@@ -276,11 +276,12 @@ pub async fn record_compliance_history(pool: &SqlitePool) -> Result<(), sqlx::Er
         .await?;
 
         let systems_score  = sys_stats.try_get::<f64, _>("avg_score").unwrap_or(0.0);
-        let systems_score = (systems_score * 100.0).round() / 100.0; 
+        let systems_score = (systems_score * 100.0).round() / 100.0;
         let policies_score = pol_stats.try_get::<f64, _>("avg_score").unwrap_or(0.0);
-        let policies_score = (policies_score * 100.0).round() / 100.0; 
-        let total_systems  = sys_stats.try_get::<i32, _>("total").unwrap_or(0);
-        let total_policies = pol_stats.try_get::<i32, _>("total").unwrap_or(0);
+        let policies_score = (policies_score * 100.0).round() / 100.0;
+        // COUNT(*) returns i64 in SQLite — use i64 to avoid overflow
+        let total_systems  = sys_stats.try_get::<i64, _>("total").unwrap_or(0);
+        let total_policies = pol_stats.try_get::<i64, _>("total").unwrap_or(0);
 
         sqlx::query(r#"
             INSERT INTO compliance_history (
